+++ src/lib/repository-factory.ts
@@ -1,4 +1,4 @@
-import { Prisma } from '@prisma/client';
+import type { Prisma } from '@prisma/client';
 import { BaseRepository } from './base-repository';
 
 export const defineRepository = <M extends Prisma.ModelName>(modelName: M) => {